refact: implement clean arch concepts

+++ src/main/kotlin/factories/CustomerFactory.kt
@@ -5,7 +5,7 @@ import domain.entities.Position
 
 object CustomerFactory {
     fun fromCsvRow(row: List<String>): Customer {
-        val id = row[0].toInt()
+        val id = row[0]
         val x = row[1].toInt()
         val y = row[2].toInt()
